function_app/requirements.txt
azure-functions
msal
requests
openpyxl
